[update] split up server files
